+++ common/src/main/java/net/satisfy/farm_and_charm/core/compat/rei/cooking/CookingPotCategory.java
@@ -9,11 +9,11 @@ import me.shedaniel.rei.api.client.registry.display.DisplayCategory;
 import me.shedaniel.rei.api.common.category.CategoryIdentifier;
 import me.shedaniel.rei.api.common.util.EntryStacks;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.satisfy.farm_and_charm.core.block.entity.CookingPotBlockEntity;
 import net.satisfy.farm_and_charm.core.registry.ObjectRegistry;
+import org.joml.Vector2i;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class CookingPotCategory implements DisplayCategory<CookingPotDisplay> {
@@ -36,8 +36,8 @@ public class CookingPotCategory implements DisplayCategory<CookingPotDisplay> {
 
     @Override
     public List<Widget> setupDisplay(CookingPotDisplay display, Rectangle bounds) {
-        var startPoint = new org.joml.Vector2i(bounds.getCenterX() - 55, bounds.getCenterY() - 13);
-        var widgets = new java.util.ArrayList<Widget>();
+        var startPoint = new Vector2i(bounds.getCenterX() - 55, bounds.getCenterY() - 13);
+        var widgets = new ArrayList<Widget>();
         widgets.add(Widgets.createRecipeBase(bounds));
         widgets.add(Widgets.createArrow(new Point(startPoint.x + 54, startPoint.y - 1))
                 .animationDurationTicks(CookingPotBlockEntity.getMaxCookingTime()));
@@ -48,8 +48,7 @@ public class CookingPotCategory implements DisplayCategory<CookingPotDisplay> {
                 .markOutput());
 
         var containerEntry = display.getInputEntries().get(6);
-        var containerStack = containerEntry.isEmpty() ? ItemStack.EMPTY : (ItemStack) containerEntry.get(0).getValue();
-        if (!containerStack.isEmpty() && containerStack.getItem() != Items.AIR) {
+        if (!containerEntry.isEmpty()) {
             widgets.add(Widgets.createSlot(new Point(startPoint.x + 56, startPoint.y + 23))
                     .entries(containerEntry)
                     .markInput());