+++ common/src/main/java/net/satisfy/farm_and_charm/core/compat/rei/doughing/CraftingBowlDisplay.java
@@ -4,7 +4,6 @@ import me.shedaniel.rei.api.common.category.CategoryIdentifier;
 import me.shedaniel.rei.api.common.display.basic.BasicDisplay;
 import me.shedaniel.rei.api.common.entry.EntryIngredient;
 import me.shedaniel.rei.api.common.util.EntryIngredients;
-import net.minecraft.resources.ResourceLocation;
 import net.satisfy.farm_and_charm.core.recipe.CraftingBowlRecipe;
 
 import java.util.ArrayList;
@@ -12,19 +11,22 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-@SuppressWarnings("all")
 public class CraftingBowlDisplay extends BasicDisplay {
     public CraftingBowlDisplay(CraftingBowlRecipe recipe) {
-        this(EntryIngredients.ofIngredients(new ArrayList<>(recipe.getIngredients())), Collections.singletonList(EntryIngredients.of(recipe.getResultItem(BasicDisplay.registryAccess()))), Optional.of(recipe.getId()));
+        super(createInputs(recipe), Collections.singletonList(EntryIngredients.of(recipe.getResultItem(BasicDisplay.registryAccess()))), Optional.of(recipe.getId()));
     }
 
-    public CraftingBowlDisplay(List<EntryIngredient> inputs, List<EntryIngredient> outputs, Optional<ResourceLocation> location) {
-        super(inputs, outputs, location);
+    private static List<EntryIngredient> createInputs(CraftingBowlRecipe recipe) {
+        List<EntryIngredient> inputs = new ArrayList<>();
+        int max = Math.min(4, recipe.getIngredients().size());
+        for (int i = 0; i < max; i++) {
+            inputs.add(EntryIngredients.ofIngredient(recipe.getIngredients().get(i)));
+        }
+        return inputs;
     }
 
     @Override
     public CategoryIdentifier<?> getCategoryIdentifier() {
         return CraftingBowlCategory.CRAFTING_BOWL_DISPLAY;
     }
-
 }