+++ common/src/main/java/net/satisfy/farm_and_charm/core/compat/rei/mincing/MincingDisplay.java
@@ -10,6 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 public class MincingDisplay extends BasicDisplay {
+
     public static final CategoryIdentifier<MincingDisplay> MINCING_DISPLAY =
             CategoryIdentifier.of(FarmAndCharm.MOD_ID, "mincing_display");